added sorting of extract blocks
