+++ src/utils/statistic.js
@@ -1,6 +1,6 @@
-import {getUsersRank} from "../utils/common.js";
-import {getObjectElementsGenres, getTopGenre} from "../utils/get-genres.js";
-import {GENRES} from './film.js';
+import {getUsersRank} from "./common.js";
+import {getObjectElementsGenres, getTopGenre} from "./get-genres.js";
+import {GENRES} from '../mock/film.js';
 
 
 const generateStatistics = (films) => {
